\docType{package}
\name{dataWKMSYREF2-package}
\alias{dataWKMSYREF2}
\alias{dataWKMSYREF2-package}
\alias{dataWKMSYREF2Library}
\title{Title dataWKMSYREF2 package}
\description{
  Contains FLStock data that were used at the WKMSYREF2
  meeting
}
\details{
  Version 1.0: 13. January 2013
}
\author{
  Various
}
\seealso{
  link to other documents
}

